+++ ha-portal/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.5
+
+- Package v1.0.1
+- Fix dark mode toggle
+- Remove missing favicon error dump from console
+
+## 1.0.4
+
+- Improve error handling and log error responses from routes
+
 ## 1.0.3
 
 - Add error store to handle top level load errors